Refactor transaction sending and package structure
