+++ packages/grill/src/index.ts
@@ -2,3 +2,4 @@ export * from "./contexts/index.js";
 export * from "./hooks/index.js";
 export * from "./providers/index.js";
 export * from "./types.js";
+export * from "./utils/index.js";
